Safari-611.1.10/LayoutTests/fast/dom/Window/window-postmessage-clone-deep-array.html

<html>
<head></head>
<body>
<div id="description"></div>
<input type="file" id="fileInput" multiple></input>
<div id="console"></div>
<script src="resources/postmessage-test.js"></script>
<script>
document.getElementById("description").innerHTML = "Tests that we support cloning deep(ish) arrays.";
deepArray=[];
for (var i = 0; i < 10000; i++)
    deepArray=[deepArray];
tryPostMessage('deepArray');
tryPostMessage('"done"');
</script>
</body>
</html>
