safari-600.7-branch/LayoutTests/platform/ios-sim/fast/dom/Window/window-postmessage-clone-deep-array-expected.txt

CONSOLE MESSAGE: line 42: RangeError: Maximum call stack size exceeded.
Tests that we support cloning deep(ish) arrays.

PASS: eventData is done of type string
